Remove global leaderboard script & Add leaderboard scene
