+++ Levels/Leaderboard.tscn
@@ -1,0 +1,112 @@
+[gd_scene load_steps=8 format=2]
+
+[ext_resource path="res://Assets/Themes/Classic311/Classic311.tres" type="Theme" id=1]
+[ext_resource path="res://Assets/Sprites/leaderboard.png" type="Texture" id=2]
+[ext_resource path="res://Assets/Jeebz-Regular.ttf" type="DynamicFontData" id=3]
+[ext_resource path="res://Levels/Leaderboard.cs" type="Script" id=5]
+[ext_resource path="res://Assets/Jeebz-Bold.ttf" type="DynamicFontData" id=6]
+
+[sub_resource type="DynamicFont" id=1]
+font_data = ExtResource( 3 )
+
+[sub_resource type="DynamicFont" id=2]
+size = 7
+font_data = ExtResource( 6 )
+
+[node name="Leaderboard" type="Control"]
+anchor_right = 1.0
+margin_left = -0.652466
+margin_top = 0.529732
+margin_right = -0.652466
+margin_bottom = 64.5297
+theme = ExtResource( 1 )
+script = ExtResource( 5 )
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="Label" type="Label" parent="."]
+margin_left = 108.253
+margin_top = -0.374565
+margin_right = 217.253
+margin_bottom = 18.6254
+custom_fonts/font = SubResource( 1 )
+custom_colors/font_color = Color( 1, 1, 1, 1 )
+text = "Leaderboard"
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="TextureRect" type="TextureRect" parent="."]
+anchor_left = 0.5
+anchor_right = 0.5
+margin_left = -71.3535
+margin_top = 14.0
+margin_right = 82.6465
+margin_bottom = 194.0
+rect_scale = Vector2( 0.952, 1 )
+texture = ExtResource( 2 )
+stretch_mode = 1
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="Panel" type="Panel" parent="."]
+margin_left = 112.0
+margin_top = 42.0
+margin_right = 210.0
+margin_bottom = 160.0
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="ScrollContainer" type="ScrollContainer" parent="Panel"]
+margin_left = 2.0
+margin_top = 2.0
+margin_right = 98.0
+margin_bottom = 117.0
+size_flags_horizontal = 3
+size_flags_vertical = 3
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="VBoxContainer" type="VBoxContainer" parent="Panel/ScrollContainer"]
+margin_right = 96.0
+size_flags_horizontal = 3
+custom_constants/separation = 0
+
+[node name="HBoxContainer" type="HBoxContainer" parent="."]
+anchor_left = 0.5
+anchor_top = 0.5
+anchor_right = 0.5
+anchor_bottom = 0.5
+margin_left = -47.0
+margin_top = 2.00002
+margin_right = 44.0
+margin_bottom = 11.0
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="Name" type="Label" parent="HBoxContainer"]
+margin_right = 43.0
+margin_bottom = 9.0
+size_flags_horizontal = 3
+custom_fonts/font = SubResource( 2 )
+text = "Name"
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="Score" type="Label" parent="HBoxContainer"]
+margin_left = 47.0
+margin_right = 91.0
+margin_bottom = 9.0
+size_flags_horizontal = 3
+custom_fonts/font = SubResource( 2 )
+text = "Score"
+align = 2
+__meta__ = {
+"_edit_use_anchors_": false
+}
