+++ Models/ScoreEntry.cs
@@ -2,10 +2,9 @@
 
 namespace Debugmancer.Models
 {
-	class ScoreEntry
+	public class ScoreEntry
 	{
 		public string Name { get; set; }
 		public int Score { get; set; }
-		public DateTime Date { get; set; }
 	}
 }
