+++ Objects/LeaderboardItem.tscn
@@ -1,0 +1,45 @@
+[gd_scene load_steps=4 format=2]
+
+[ext_resource path="res://Assets/Jeebz-Regular.ttf" type="DynamicFontData" id=1]
+[ext_resource path="res://Assets/Sprites/item.png" type="Texture" id=2]
+
+[sub_resource type="DynamicFont" id=1]
+size = 5
+font_data = ExtResource( 1 )
+
+[node name="Item" type="TextureRect"]
+margin_right = 97.0
+margin_bottom = 16.0
+texture = ExtResource( 2 )
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="HBoxContainer" type="HBoxContainer" parent="."]
+margin_left = 3.0
+margin_top = 2.0
+margin_right = 90.0
+margin_bottom = 16.0
+__meta__ = {
+"_edit_use_anchors_": false
+}
+
+[node name="Name" type="Label" parent="HBoxContainer"]
+margin_top = 3.0
+margin_right = 41.0
+margin_bottom = 10.0
+size_flags_horizontal = 3
+custom_fonts/font = SubResource( 1 )
+custom_colors/font_color = Color( 0.176471, 0.176471, 0.176471, 1 )
+text = "user#0000"
+
+[node name="Score" type="Label" parent="HBoxContainer"]
+margin_left = 45.0
+margin_top = 3.0
+margin_right = 87.0
+margin_bottom = 10.0
+size_flags_horizontal = 3
+custom_fonts/font = SubResource( 1 )
+custom_colors/font_color = Color( 0.176471, 0.176471, 0.176471, 1 )
+text = "1234"
+align = 2
